+++ frontend/src/components/InputIngredient/index.ts
